Package wide refactoring to update the embedders
