+++ embiggen/embedder/word2vec/callbacks/callback.py
@@ -1,0 +1,4 @@
+class Callback:
+
+    def __init__(self):
+        pass
