+++ embiggen/embedder/word2vec/callbacks/history.py
@@ -1,0 +1,6 @@
+from .callback import Callback
+
+class History(Callback):
+
+    def __init__(self):
+        pass
