+++ embiggen/graph/graph.py
@@ -103,6 +103,96 @@ class Graph(Hashable):
         """Return mapping for numeric nodes ID to nodes."""
         return self.nodes_indices
 
+    # !TODO: Integrate this with graph!
+    def DegreeProduct(graph: Graph, node_1, node_2):
+        ''' Function takes a CSF graph object and list of edges calculates the Degree Product or Preferential Attachment for these
+        nodes given the structure of the graph.
+        :param graph: Graph  object
+        :param node_1: one node of graph
+        :param node_2: one node of a graph
+        :return: Degree Product score of the nodes
+        '''
+
+        return graph.degree(node_1) * graph.degree(node_2)
+
+    # !TODO: Integrate this with graph!
+    def CommonNeighbors(graph: Graph, node_1, node_2):
+        ''' Function takes a CSF graph object and list of edges calculates the Common Neighbors for these nodes given the
+        structure of the graph.
+        :param graph: Graph object
+        :param node_1: one node of graph
+        :param node_2: one node of a graph
+        :return: Common Neighbors score of the nodes
+        '''
+
+        node_1_neighbors = set(graph.neighbors(node_1))
+        node_2_neighbors = set(graph.neighbors(node_2))
+
+        if len(node_1_neighbors) == 0 or len(node_2_neighbors) == 0:
+            score = 0.0
+
+        elif len(node_1_neighbors.intersection(node_2_neighbors)) == 0:
+            score = 0.0
+
+        else:
+            n_intersection = node_1_neighbors.intersection(node_2_neighbors)
+            score = float(len(n_intersection))
+
+        return score
+
+    # !TODO: Integrate this with graph!
+    def Jaccard(graph: Graph, node_1, node_2):
+        ''' Function takes a CFS graph object and list of edges calculates the Jaccard for these nodes given the
+        structure of the graph.
+        :param graph: CFS graph object
+        :param node_1: one node of graph
+        :param node_2: one node of a graph
+        :return: The Jaccad score of two nodes
+        '''
+
+        node_1_neighbors = set(graph.neighbors(node_1))
+        node_2_neighbors = set(graph.neighbors(node_2))
+
+        if len(node_1_neighbors) == 0 or len(node_2_neighbors) == 0:
+            score = 0.0
+
+        elif len(node_1_neighbors.intersection(node_2_neighbors)) == 0:
+            score = 0.0
+
+        else:
+            n_intersection = set(node_1_neighbors.intersection(node_2_neighbors))
+            n_union = set(node_1_neighbors.union(node_2_neighbors))
+            score = float(len(n_intersection))/len(n_union)
+
+        return score
+
+    # !TODO: Integrate this with graph!
+    def AdamicAdar(graph: Graph, node_1, node_2):
+        ''' Function takes a CSF graph object and list of edges calculates the Adamic Adar for the nodes given the
+        structure of the graph.
+        :param graph: CSF graph object
+        :param node_1: a node of the graph
+        :param node_2: one node of a graph
+        :return: AdamicAdar score of the nodes
+        '''
+
+        node_1_neighbors = set(graph.neighbors(node_1))
+        node_2_neighbors = set(graph.neighbors(node_2))
+
+        if len(node_1_neighbors) == 0 or len(node_2_neighbors) == 0:
+            score = 0.0
+
+        elif len(node_1_neighbors.intersection(node_2_neighbors)) == 0:
+            score = 0.0
+
+        else:
+            score = 0.0
+            n_intersection = node_1_neighbors.intersection(node_2_neighbors)
+
+            for c in n_intersection:
+                score += 1/np.log(graph.degree(c))
+        return score
+
     def degree(self, node: int) -> int:
         """Return degree of given node.
