--- embiggen/w2v/__init__.py
@@ -1,4 +1,0 @@
-from .cbow_list_batcher import CBOWListBatcher
-from .skip_gram_batcher import SkipGramBatcher
-
-__all__ = ["CBOWListBatcher", "SkipGramBatcher"]
